workers/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
celery[redis]==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
beautifulsoup4==4.12.3
requests==2.32.3
feedparser==6.0.11
pytrends==4.9.2
praw==7.8.1
python-dotenv==1.0.1
httpx==0.28.1
lxml==5.3.0
openai==1.59.6
